# lib/api-zod/src/generated/types/inquirySubmission.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rehman INC API - Real estate investment company lead management
 * OpenAPI spec version: 0.1.0
 */
import type { InquirySubmissionPreferredContact } from './inquirySubmissionPreferredContact';
import type { InquirySubmissionPropertyCondition } from './inquirySubmissionPropertyCondition';
import type { InquirySubmissionPropertyType } from './inquirySubmissionPropertyType';
import type { InquirySubmissionSellingTimeline } from './inquirySubmissionSellingTimeline';

export interface InquirySubmission {
  /** @minLength 3 */
  address: string;
  /** @minLength 1 */
  city: string;
  /** @minLength 2 */
  state: string;
  /** @minLength 5 */
  zip: string;
  /** @nullable */
  propertyType?: InquirySubmissionPropertyType;
  /** @nullable */
  bedrooms?: string | null;
  /** @nullable */
  bathrooms?: string | null;
  /** @nullable */
  squareFootage?: string | null;
  /** @nullable */
  occupied?: string | null;
  /** @nullable */
  propertyCondition?: InquirySubmissionPropertyCondition;
  /** @nullable */
  repairs?: string | null;
  /** @nullable */
  sellingReason?: string | null;
  /** @nullable */
  sellingTimeline?: InquirySubmissionSellingTimeline;
  /** R2 object keys for uploaded photos */
  photoKeys?: string[];
  /** @minLength 2 */
  fullName: string;
  email: string;
  /** @minLength 10 */
  phone: string;
  preferredContact: InquirySubmissionPreferredContact;
  contactConsent: boolean;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  utmSource?: string | null;
  /** @nullable */
  utmMedium?: string | null;
  /** @nullable */
  utmCampaign?: string | null;
}