# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rehman INC API - Real estate investment company lead management
 * OpenAPI spec version: 0.1.0
 */

export interface SuccessResponse {
  success: boolean;
}